requirements.txt
faiss_cpu==1.9.0
fastapi==0.128.4
feedparser==6.0.11
langchain==1.2.9
langchain_community==0.4.1
langchain_core==1.2.9
langchain_ollama==1.0.1
langchain_openai==1.1.7
langgraph==1.0.8
numpy==2.4.2
pdfplumber==0.11.4
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.0
uvicorn
langchain_google_genai
langchain_text_splitters
